Move classes into package and fix imports
