+++ S3Scanner/exceptions.py
